- FUntypedBulkData::SetBulkDataAlignment now takes a uint16 as that is the storage type used anyway.
- FBulkDataBuffer now uses TArrayView64 as the wrapper around it's data. This has been requested by another team anyway.
-- SoundWave.cpp will pass on the return type of FBulkDataBuffer to other systems. So I moved the 'is the data too big for TArrayView checks there and did not change the SoundWave api.

#rb Per.Larsson, Max.Hayes
#rnx
#preflight 60a606414c320300018014f4

#ROBOMERGE-SOURCE: CL 16399515 in //UE5/Main/...
#ROBOMERGE-BOT: STARSHIP (Main -> Release-Engine-Test) (v804-16311228)

[CL 16399762 by paul chipchase in ue5-release-engine-test branch]

+++ Engine/Source/Runtime/CoreUObject/Private/Serialization/BulkData.cpp
@@ -831,7 +831,7 @@ uint32 FUntypedBulkData::GetBulkDataFlags() const
  *
  * @param BulkDataAlignmentToSet	Bulk data alignment to set
  */
-void FUntypedBulkData::SetBulkDataAlignment( uint32 BulkDataAlignmentToSet )
+void FUntypedBulkData::SetBulkDataAlignment(uint16 BulkDataAlignmentToSet)
 {
 	BulkDataAlignment = BulkDataAlignmentToSet;
 }
@@ -2376,7 +2376,7 @@ void FFloatBulkDataOld::SerializeElement( FArchive& Ar, void* Data, int64 Elemen
 	Ar << FloatData;
 }
 
-void FFormatContainer::Serialize(FArchive& Ar, UObject* Owner, const TArray<FName>* FormatsToSave, bool bSingleUse, uint32 InAlignment, bool bInline, bool bMapped)
+void FFormatContainer::Serialize(FArchive& Ar, UObject* Owner, const TArray<FName>* FormatsToSave, bool bSingleUse, uint16 InAlignment, bool bInline, bool bMapped)
 {
 	if (Ar.IsLoading())
 	{
